# Change BuildInput to return (BuildLabel, bool) instead of *BuildLabel