--- a/src/core/build_env.go
+++ b/src/core/build_env.go
@@ -313,12 +313,12 @@ func initStampEnv() {
 }
 
 func toolPath(state *BuildState, tool BuildInput, abs bool) string {
-	if label := tool.Label(); label != nil {
+	if label, ok := tool.Label(); ok {
 		entryPoint := ""
 		if o, ok := tool.(AnnotatedOutputLabel); ok {
 			entryPoint = o.Annotation
 		}
-		path := state.Graph.TargetOrDie(*label).toolPath(abs, entryPoint)
+		path := state.Graph.TargetOrDie(label).toolPath(abs, entryPoint)
 		if !strings.Contains(path, "/") {
 			path = "./" + path
 		}