--- a/src/core/build_target_test.go
+++ b/src/core/build_target_test.go
@@ -903,7 +903,9 @@ func TestIsTool(t *testing.T) {
 	target.AddTool(withEP)
 
 	assert.True(t, target.IsTool(noEP))
-	assert.True(t, target.IsTool(*withEP.Label()))
+	l, ok := withEP.Label()
+	assert.True(t, ok)
+	assert.True(t, target.IsTool(l))
 }
 
 func makeTarget1(label, visibility string, deps ...*BuildTarget) *BuildTarget {