This seems to save an allocation when called:
Before:
```
BenchmarkProvideFor
BenchmarkProvideFor/Simple
BenchmarkProvideFor/Simple-12           63322402                18.81 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/NoMatch
BenchmarkProvideFor/NoMatch-12          31253348                38.50 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/OneMatch
BenchmarkProvideFor/OneMatch-12          8090606               145.7 ns/op            48 B/op          1 allocs/op
BenchmarkProvideFor/TwoMatches
BenchmarkProvideFor/TwoMatches-12        4717500               247.8 ns/op            96 B/op          1 allocs/op
BenchmarkProvideFor/IsData
BenchmarkProvideFor/IsData-12            8401734               128.9 ns/op            48 B/op          1 allocs/op
```
After:
```
BenchmarkProvideFor
BenchmarkProvideFor/Simple
BenchmarkProvideFor/Simple-12           64205665                18.43 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/NoMatch
BenchmarkProvideFor/NoMatch-12          32368802                36.85 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/OneMatch
BenchmarkProvideFor/OneMatch-12          6848133               152.0 ns/op            48 B/op          1 allocs/op
BenchmarkProvideFor/TwoMatches
BenchmarkProvideFor/TwoMatches-12        4521673               249.3 ns/op            96 B/op          1 allocs/op
BenchmarkProvideFor/IsData
BenchmarkProvideFor/IsData-12           36188140                31.77 ns/op            0 B/op          0 allocs/op
```
i.e. we save an alloc in the last path where we call `.Label()`. 

This obviously touches a lot of files but should be a fairly straightforward refactor (and that suggests there are a bunch more places that will get mild benefits from this too).